rename notification to notifier
based on @excid3 recommendation in https://github.com/excid3/noticed/pull/258#issuecomment-1349585670
