+++ lib/generators/noticed/templates/notifier.rb.tt
